restructuring
after using this framework on multiple projects, I'm restructuring many of the files and adding in more conditional variables to turn swatches of smaller code bits on / off.

this is not a 'complete' commit.  need to take another pass through and finish any more conversions and new variable base values for future setup file